+++ azure/eventhub/async_ops/sender_async.py
@@ -97,8 +97,8 @@ class AsyncSender(Sender):
                 properties=self.client.create_properties(),
                 loop=self.loop)
         await self._handler.open_async()
-        while not await self.has_started():
-            await self._handler._connection.work_async()  # pylint: disable=protected-access
+        while not await self._handler.client_ready_async():
+            await asyncio.sleep(0.05)
 
     async def reconnect_async(self):
         """If the Receiver was disconnected from the service with
@@ -148,6 +148,7 @@ class AsyncSender(Sender):
         Whether the handler has completed all start up processes such as
         establishing the connection, session, link and authentication, and
         is not ready to process messages.
+        **This function is now deprecated and will be removed in v2.0+.**
 
         :rtype: bool
         """