+++ azure/eventhub/receiver.py
@@ -2,9 +2,11 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
+from __future__ import unicode_literals
 
 import uuid
 import logging
+import time
 
 from uamqp import types, errors
 from uamqp import ReceiveClient, Source
@@ -14,7 +16,7 @@ from azure.eventhub.common import EventHubError, EventData, _error_handler
 log = logging.getLogger(__name__)
 
 
-class Receiver:
+class Receiver(object):
     """
     Implements a Receiver.
     """
@@ -97,8 +99,8 @@ class Receiver:
                 client_name=self.name,
                 properties=self.client.create_properties())
         self._handler.open()
-        while not self.has_started():
-            self._handler._connection.work()
+        while not self._handler.client_ready():
+            time.sleep(0.05)
 
     def reconnect(self):
         """If the Receiver was disconnected from the service with
@@ -106,7 +108,7 @@ class Receiver:
         # pylint: disable=protected-access
         alt_creds = {
             "username": self.client._auth_config.get("iot_username"),
-            "password":self.client._auth_config.get("iot_password")}
+            "password": self.client._auth_config.get("iot_password")}
         self._handler.close()
         source = Source(self.source)
         if self.offset is not None:
@@ -124,8 +126,8 @@ class Receiver:
             properties=self.client.create_properties())
         try:
             self._handler.open()
-            while not self.has_started():
-                self._handler._connection.work()
+            while not self._handler.client_ready():
+                time.sleep(0.05)
         except (errors.LinkDetach, errors.ConnectionClose) as shutdown:
             if shutdown.action.retry and self.auto_reconnect:
                 self.reconnect()
@@ -160,6 +162,7 @@ class Receiver:
         Whether the handler has completed all start up processes such as
         establishing the connection, session, link and authentication, and
         is not ready to process messages.
+        **This function is now deprecated and will be removed in v2.0+.**
 
         :rtype: bool
         """